rename BFF projects/templates
