+++ Duende.IdentityServer.Templates.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31112.23
+# Visual Studio Version 17
+VisualStudioVersion = 17.6.33717.318
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IdentityServerEmpty", "src\IdentityServerEmpty\IdentityServerEmpty.csproj", "{D324CA74-877A-43AB-B978-38416BBE300A}"
 EndProject
@@ -11,9 +11,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IdentityServerAspNetIdentit
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IdentityServerEntityFramework", "src\IdentityServerEntityFramework\IdentityServerEntityFramework.csproj", "{46000F46-B039-49F3-AACD-42E4FBC11E94}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BffJs", "src\BffJs\BffJs.csproj", "{D471E5B4-1D16-4861-81B9-5DD4DC0786BA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BffLocalApi", "src\BffLocalApi\BffLocalApi.csproj", "{B0948334-5301-41AF-864E-03D335E766B3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GenericBFF", "src\GenericBFF\GenericBFF.csproj", "{E9A70AFB-7B75-46BF-BB47-F00B63D22618}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BffRemoteApi", "src\BffRemoteApi\BffRemoteApi.csproj", "{B1E3EDAC-9C5A-4D3B-A6D2-2A1AC0D16728}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -37,14 +37,14 @@ Global
 		{46000F46-B039-49F3-AACD-42E4FBC11E94}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{46000F46-B039-49F3-AACD-42E4FBC11E94}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{46000F46-B039-49F3-AACD-42E4FBC11E94}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D471E5B4-1D16-4861-81B9-5DD4DC0786BA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D471E5B4-1D16-4861-81B9-5DD4DC0786BA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D471E5B4-1D16-4861-81B9-5DD4DC0786BA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D471E5B4-1D16-4861-81B9-5DD4DC0786BA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E9A70AFB-7B75-46BF-BB47-F00B63D22618}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E9A70AFB-7B75-46BF-BB47-F00B63D22618}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E9A70AFB-7B75-46BF-BB47-F00B63D22618}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E9A70AFB-7B75-46BF-BB47-F00B63D22618}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B0948334-5301-41AF-864E-03D335E766B3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B0948334-5301-41AF-864E-03D335E766B3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B0948334-5301-41AF-864E-03D335E766B3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B0948334-5301-41AF-864E-03D335E766B3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B1E3EDAC-9C5A-4D3B-A6D2-2A1AC0D16728}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B1E3EDAC-9C5A-4D3B-A6D2-2A1AC0D16728}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B1E3EDAC-9C5A-4D3B-A6D2-2A1AC0D16728}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B1E3EDAC-9C5A-4D3B-A6D2-2A1AC0D16728}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
