--- src/BffJs/.template.config/template.json
@@ -1,19 +1,0 @@
-{
-  "author": "Duende Software",
-  "classifications": [
-    "Web",
-    "IdentityServer"
-  ],
-  "name": "Duende BFF with JavaScript",
-  "identity": "IdentityServer.BFF.JavaScript",
-  "groupIdentity": "IdentityServer.BFF.JavaScript",
-  "shortName": "bffjs",
-  "tags": {
-    "language": "C#",
-    "type": "project"
-  },
-
-  "sourceName": "BffJs",
-  "preferNameDirectory": "true",
-  "primaryOutputs": [ { "path": "BffJs.csproj" } ]
-}
