+++ src/BffLocalApi/.template.config/template.json
@@ -1,0 +1,19 @@
+{
+  "author": "Duende Software",
+  "classifications": [
+    "Web",
+    "IdentityServer"
+  ],
+  "name": "Duende BFF using a Local API",
+  "identity": "IdentityServer.BFF.LocalApi",
+  "groupIdentity": "IdentityServer.BFF.LocalApi",
+  "shortName": "bff-localapi",
+  "tags": {
+    "language": "C#",
+    "type": "project"
+  },
+
+  "sourceName": "BffLocalApi",
+  "preferNameDirectory": "true",
+  "primaryOutputs": [ { "path": "BffLocalApi.csproj" } ]
+}
