+++ src/BffRemoteApi/.template.config/template.json
@@ -1,0 +1,23 @@
+{
+  "author": "Duende Software",
+  "classifications": [
+    "Web",
+    "IdentityServer"
+  ],
+  "name": "Duende BFF Host using a Remote API",
+  "identity": "IdentityServer.BFF.RemoteApi",
+  "groupIdentity": "IdentityServer.BFF.RemoteApi",
+  "shortName": "bff-remoteapi",
+  "tags": {
+    "language": "C#",
+    "type": "project"
+  },
+  "sourceName": "BffRemoteApi",
+  "preferNameDirectory": "true",
+  "primaryOutputs": [
+    {
+      "path": "BffRemoteApi.csproj"
+    }
+  ]
+}
+}
