+++ src/BffRemoteApi/wwwroot/app.js
